fixed segfault

+++ zkboo/prover.h
@@ -42,7 +42,7 @@ class Proof {
         CircuitView *views[2];
         RandomSource rands[2];
         uint8_t *w[2];
-        uint8_t *outShares[2];
+        uint32_t *outShares[2];
         uint8_t *out;
         int output;
         int idx;
